feat: link EHR and DICOMWeb using PatientId
Link the EHR records with DICOMWeb fetched images using PatientId.
EHR will be the driving database. When user selects a patient
from the EHR list of patients, the DICOMWeb server will be queried
to fetch all studies associated with that patient.
DICOMWeb is not queried until a patient is selected.