adapt lungair_methods.py for new VolView python server design
The server/methods/ directory used to be built into core volview, and our lungair-main branch when we forked core volview relied on this directory being there.
Now that after the migration, this directory is gone from the more recent core volview that we are pinning our submodule to.
We can't go back to an earlier volview because the python server didn't exist when we forked.
I copied from this newer example to updated how lungair_methods.py works:
https://github.com/Kitware/VolView/blob/411e5a891bfb520647ab3f97cac6edfcca930a65/server/examples/example_api.py

Fix lungair_methods.py to work with the newer VolView python server design